requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
python-box
streamlit
python-dotenv
replicate
transformers
replicate
